examples/mobc: apply rename migration(System -> system)
