+++ examples/mobc/src/src_user/Settings/Modes/Transitions/sl_nop.c
@@ -4,7 +4,7 @@
 #include "../../../TlmCmd/command_definitions.h"
 
 #include <src_core/TlmCmd/block_command_loader.h>
-#include <src_core/System/TimeManager/obc_time.h>
+#include <src_core/system/time_manager/obc_time.h>
 
 void BCL_load_sl_nop(void)
 {
