+++ examples/mobc/src/src_user/Settings/system/app_manager_params.h
