+++ examples/mobc/src/src_user/Settings/system/event_handler_rules/event_handler_rule_test.c
@@ -4,7 +4,7 @@
  * @brief  EH の テスト用 Rule
  */
 #include "event_handler_rules.h"
-#include <src_core/System/EventManager/event_handler.h>
+#include <src_core/system/event_manager/event_handler.h>
 #include "../../../TlmCmd/block_command_definitions.h"
 
 void EH_load_rule_test(void)
