+++ examples/mobc/src/src_user/Settings/system/event_handler_rules/event_handler_rules.c
