+++ examples/mobc/src/src_user/TlmCmd/command_source.h
@@ -5,13 +5,13 @@
 #ifndef COMMAND_SOURCE_H_
 #define COMMAND_SOURCE_H_
 
-#include <src_core/System/EventManager/event_logger.h>
-#include <src_core/System/EventManager/event_handler.h>
-#include <src_core/System/TimeManager/time_manager.h>
-#include <src_core/System/ModeManager/mode_manager.h>
-#include <src_core/System/ApplicationManager/app_manager.h>
-#include <src_core/System/TaskManager/task_dispatcher.h>
-#include <src_core/System/WatchdogTimer/watchdog_timer.h>
+#include <src_core/system/event_manager/event_logger.h>
+#include <src_core/system/event_manager/event_handler.h>
+#include <src_core/system/time_manager/time_manager.h>
+#include <src_core/system/mode_manager/mode_manager.h>
+#include <src_core/system/application_manager/app_manager.h>
+#include <src_core/system/task_manager/task_dispatcher.h>
+#include <src_core/system/watchdog_timer/watchdog_timer.h>
 #include <src_core/TlmCmd/block_command_executor.h>
 #include <src_user/TlmCmd/block_command_user_settings.h>
 #include <src_core/TlmCmd/command_dispatcher_manager.h>
