+++ examples/mobc/src/src_user/applications/driver_instances/di_uart_test.h
@@ -6,7 +6,7 @@
 #define DI_UART_TEST_H_
 
 #include "../../Drivers/Etc/uart_test.h"
-#include <src_core/System/ApplicationManager/app_info.h>
+#include <src_core/system/application_manager/app_info.h>
 #include <src_core/TlmCmd/common_cmd_packet.h>
 
 extern const UART_TEST_Driver* const uart_test_driver;
