+++ examples/mobc/src/src_user/applications/user_defined/debug_apps.h
@@ -1,7 +1,7 @@
 #ifndef DEBUG_APPS_H_
 #define DEBUG_APPS_H_
 
-#include <src_core/System/ApplicationManager/app_info.h>
+#include <src_core/system/application_manager/app_info.h>
 
 AppInfo APP_DBG_flush_screen(void);
 AppInfo APP_DBG_print_time_stamp(void);
